Add support for signing files
Add support for signing files.

If a file has been delay-signed and we're real-signing the build, we
create a <FilesToSign> item pointing at the output binary. The Microsoft
build systems will use this to sign the appropriate binaries at the end
of a project build.

Note that this currently only supports signing .dll and .exe binaries;
it does not yet support signing .vsix files.