Correct removed api logic
The logic for detecting a removed API in the unshipped API text was
comparing against the full entry.  This meant it compared the calculated
name to the stored name + the removed prefix.  Updated to compare
against the correct name.

There was a unit test to catch this case but the unit tests themselves
weren't functioning correctly.  They didn't include the error because it
was in an additional document.  Updated the unit test to report this
type of error.