Fix up unit tests
Changes:
1. Fix whitespace. The fixer tests had a mix of tabs and spaces,
which did not appear to be intentional. This sometimes caused the
equality check to fail when there were no substantive differences.
All of the tabs have been converted to spaces.
2. Fix expected output. In a couple of cases the expected output
after the fix was the same as the input; presumably it was never
updated because the test couldn't succeed until the underlying
issue was addressed, anyway.

Related to #285.