Add project-specific docs
Add project-specific documentation. At the moment this documentation
is just a list of diagnostics produced by a set of analyzers with
some basic information--ID, title, description, etc. A simple tool
was created to produce this documentation in Markdown format.

The main README.md has been updated to point to this information.