Add a tool to generate analyzer documentation
Add a tool to generate analyzer documentation in Markdown format.
a2md loads the assemblies specified on the command line, retrieves
the analyzers from the assemblies, and then retrieves all the
`DiagnosticDescriptors` from those analyzers. It then sorts them
by ID, and outputs the ID, title, description, category, severity,
and help link to standard output in Markdown format.

This commit also includes a script file, BuildAnalyzerMarkdown.bat,
that runs a2md over the analyzers in the Binaries\Debug folder and
redirects the output to the appropriate locations in the src folder.
The idea is that after adding/updating an analyzer and rebuilding you
can simply run the script and get all the updated documentation.