Allow diagnostics in external files
Right now the unit testing framework assumes a diagnostic will be in a
source file. If the diagnostic is not in a source file, we assume it is
in a metadata file and fail the test because we don't support this. This
also means we don't support diagnostics in an additional file, since it
isn't a source file.

Instead we should fail the test only if the diagnostic location is
explicitly in metadata.

Also, results in additional files must have been created with the
GetResultAt method rather than GetCSharpResultAt or GetBasicResultAt, so
update the test code to account for that situation.