Look for duplicated symbols in public API files
Currently, the DeclarePublicAPIAnalyzer does not expect the same symbol
to occur more than once, either within a public API file or across
files. If this happens the analyzer throws an exception in its
compilation start action and fails to actually check the API.

This commit updates the analyzer to look out for duplicated symbols and
create a diagnostic to bring the issue to the attention of the user.

Fixes #312.