Make DeclarePublicAPI diagnostics warnings
Change the DeclarePublicAPI diagnostics to be warnings by default,
instead of errors. This way the diagnostics will still fail pre-checkin
runs on local systems and pull request builds in the cloud (because we
turn on TreatWarningsAsErrors in both those cases) but don't get in your
way while you're actively making changes.