Fix build break in AnalyzersStatusGenerator tool
If you did "git clean -xdf" followed by cibuild.cmd, you would get
a build break in the AnalyzersStatusGenerator project, because it
couldn't resolve the reference to the assembly
System.Composition.AttributedModel.

The problem was that the .csproj file specified a HintPath for this
assembly in ..\packages, but actually the packages directory is three
levels up (..\..\..\packages). Several other assembly references in this
project file had the same problem, but those assemblies weren't actually
needed.

So I corrected the HintPath for System.Composition.AttributedModel, and
removed the unncessary assembly references. All the remaining assembly
references in this project file correctly specified ..\..\..\packages.

The odd thing is that if you built a second time, the build would succeed.
Apparently on the second attempt it didn't bother to try building this
project (that must be so, or else it would have failed again for the same
reason).