CA1716: Test nested types
Per suggestion from @nguerrera, add unit tests for nested types to ensure that nobody accidentally "simplifies" the code by replacing the usage of GetResultantVisibility with DeclaredVisibility.

Adding these tests revealed a bug, namely that in the TypeRule, I was not properly formatting the type name for nested classes; that is, I was displaying "D" instead of "C.D".