Fix #451: Partial #451: CA1716: IdentifiersShouldNotMatchKeywords (namespace names)
Conclude the odyssey by implementing the rule that namespaces names should
not match keyword names.

Per discussion with @srivatsn, we don't report a location for this rule.
The reason is that you can open a namespace in multiple locations. We
only want to report a violation for a given namespace once, so we keep
track of which namespaces we've reported. But which of the multiple
occurrences the analyzer will find first, and report, is not
deterministic.

Why does that matter? Suppose the namespace occurs in both A.cs and B.cs.
Suppose the compiler reports the violation in A.cs, and the user
suppresses it. In contrast to the way it treats partial classes, Roslyn
does _not_ suppress a namespace-scoped diagnostic across all occurrences
of the namespace. So if, in a future compilation, the compiler decides to
report the violation in B.cs, the user will see the diagnostic that he has
suppressed reappear.

I filed this as roslyn/#8643.