Add support to DeclarePublicAPI fixer to remove stale entries for sibling symbols when adding a new entry for a symbol.
One of the biggest complaints for the current Public API code fix is that changing the signature of an existing unshipped public API causes the code fix to add a new public API entry, but the old entry in the unshipped API file is not removed. This change fixes that by walking through all entries in the unshipped public API file to find entries for siblings of new member being added, and removing the ones that are no longer valid entries, i.e. either removed or are now non-public.

There were no fixer tests for public API analyzer - so I had to add them and also add test framework support for passing around additional files.

Fixes #756