CA1716: Fixes race condition in reporting namespace warnings (#1013)
This fixes #1012.

After eliminating the race condition, https://github.com/dotnet/roslyn-analyzers/blob/master/src/Microsoft.ApiDesignGuidelines.Analyzers/UnitTests/IdentifiersShouldNotMatchKeywordsNamespaceRuleTests.cs#L101 started to fail. This test expects unique warnings based on the combination of namespace and keyword. But the `HashSet` only tracked namespaces, so it reported once instead of twice.

To solve this, I've changed the `HashSet` to store the combination of namespace and keyword.