Fix to CA1003 to ignore interface implementations
This fixes issue #1087, where CA1003 misappropriately fires on event handlers for violating the rule if they implement an interface that contains a non-generic event handler. For example, if a developer wants to implement System.ComponentModel.INotifyPropertyChanged--or another interface they do not control--they cannot without firing this rule.

This rule should continue to fire where those events are being declared in an interface, but not on overrides or implementations of it. The code previously includes overrides but does not include implementations. This commit fixes now includes implementations.