Fix cibuild.cmd: assuming VS150COMNTOOLS exists (#1139)
"VS2017 does not add VS150COMNTOOLS as a system environment variable." [1] Therefore, cibuild.cmd only currently works from a VS 2017 developer prompt. This is unideal, as cibuild.cmd should load the appropriate environment, as it did in VS 2015. "This makes it easier for people to simply clone this repo and build." [2]

This is a port of Microsoft/visualfsharp@bf52776167fe6a9f2354ea96094a025191dbd3e7 [2].

[1] https://github.com/Microsoft/visualfsharp/issues/1761
[2] https://github.com/Microsoft/visualfsharp/pull/2690