Fix to CA2119 and #1086: Ignore classes that cannot be subtyped outside assembly (#1136)
* Fix to CA2119 and #1086: Ignore classes that cannot be subtyped outside assembly

By definition, this rule should not fire if the class cannot be subclassed outside the current assembly. Classes that are declared as 'internal' fit the bill, as well as classes that do not have public constructors. This change addresses these gaps.

* Clarify comment in CA2119 on internal types

Internal types can indeed be subclassed outside of the current assembly via [InternalsVisibleTo]. This change updates the comment to clarify this scenario and to confirm that it does not apply to CA2119.

* Fix to CA2119 to ignore sealed classes

Sealed classes cannot be subclassed anywhere so by definition, CA2119 does not apply to their members.