Fix CA1720 reporting violating parameters in derived types (#1095)
This fixes issue #1094 where CA1720 ignores the violating parameter name of 'obj' in Equals(object obj) and GetHashCode(T obj) but reports other violating parameter names defined in base types. This is a problem because the developer may be helpless to correct the base class/interface, and if they make the fix this rule requires, CA1725 will fire: Parameter names should match base declaration. In the end, one of the rules will have to be suppressed.

Instead, the changes in this commit prevent the parameter names in overridden methods/interface implementation methods from being checked. If the offending base class/interface is within control of the developer, then *it* will show the violation.