Add CA1061: Do not hide base class methods (#1100)
* Add CA1061: Do not hide base class methods

This code should duplicate the FxCop rule CA1061 exactly except when the base type method is private. In this circumstance, my implementation does not report a diagnostic while FxCop does.

The fixer is only boilerplate at this time.

* Update tests and correct review feedback

Renamed and reorganized unit tests so the flow is more logical, added additional cases, and created corresponding tests for Visual Basic. Then I restructured `GetMethodsHiddenByName` so that it's conceptually simpler and more efficient. It also handles that situation where parameter types are different but unrelated.

* Change severity and specify generated code analysis

* Fix/test bug related to parameter type mismatch order