Fix the VSIX break introduced from prior manifest change
We changed the minimu required VS version for FxCop analyzers VSIX from 15.0 to 15.3. However, the version of vsixinstaller.exe in Dev15.3 preview is still 15.0, and hence this blocks the VSIX installation. So, we are reverting this change.
Additionally, I also enabled our version check analyzer to be turned on by default for both VSIX and NuGet (currently it was on only for NuGet packages).