Remove usafe of IFieldInitializer in UseLiteralsWhereAppropriate analyzer
This API will be removed with https://github.com/dotnet/roslyn/issues/21293, and this change moves the analyzer to use operation blocks.
The analyzer doesn't work now as the operation block root node is still an IFieldInitializer, not the expression to the right of the equals token.
Additionally, even when https://github.com/dotnet/roslyn/issues/21293 is implemented, we will need to fix the back compat issue https://github.com/dotnet/roslyn-analyzers/issues/1300.