Handle loops and if-else statements and fix a bug in merge algorithm
Addresses #1567

Add temporary code to DataFlowOperationVisitor to handle flow analysis for loops and if-else statements. This code will be removed once we move to compiler's CFG, which would have lowered all IOperations for these constructs. #1567 will be kept active to track this code removal.
Enabling skipped tests for if-else statement identified a bug in the merge algorithm for merging analysis data in presence of points to analysis, which has now been fixed.