Improve preciseness of PointsToAnalysis when merging known and unknown values from different control paths.
Previously, if the PointsTo value was unknown in one of the control paths, Merge returned an unknown PointsTo result with zero locations. This PR changes the merge logic to keep track of subset of known locations that come from each of the control flow paths, even if the merge result has kind unknown, indicating that it can point to one of the known locations or an unknown location outside the known location set.