Only add getters and setters for type-forwarded properties to public API
During normal operation, properties are not added to the public API surface area because the analyzer does not register a symbol
action for `SymbolKind.Property`. This works fine, there is an action registered for `SymbolKind.Method`, which ensures that getters
and setters are added as expected. Things are a bit different for forwarded types.

In the case of a forwarded type, the public API surface area is discovered by recursively searching members and nested types. However,
this code adds both properties *and* accessors, resulting in the PublicAPI.Shipped.txt file needs to include an entry for the property
and the accessors, like so:

```
Microsoft.CodeAnalysis.FileTextLoader.Path.get -> string (forwarded, contained in Microsoft.CodeAnalysis.Workspaces)
Microsoft.CodeAnalysis.FileTextLoader.Path { get; } -> string (forwarded, contained in Microsoft.CodeAnalysis.Workspaces)
```

This change fixes the analyzer to not consider property symbols when recursively searching through forwarded types. However, it still
considers their accessors.