Improve wording for RS1010 diagnostic
The prior wording suggested that the EquivalenceKey has to be absolutely unique. The new wording attempts to better convey the idea that the value SHOULD be reused for every instance of CodeAction that performs a semantically similar code fix, even if applied across multiple diagnostic locations.

See https://github.com/Microsoft/vs-threading/issues/277#issuecomment-398546177 for discussion.