#1638 restrict to safe replacements, add more unit tests
the one-parameter replacement sb.Append(xxxx.Substring(a)) to sb.Append(xxxx, a, xxxx.Length - a)
is 1) dangerous when xxxx has side effects, as it's code is executed twice after the "fix",
and 2) may have performance problems.

Therefore this commit restricts to some Operation types that can be considered to be safe.
provided substrings cover at least one case per operation kind that works.
For the other OperationKinds, where no diagnostic must be reported I did not find ways yet how to create code that works reasonable.
Happy to get hints here or advice what can safely be omitted as test cases. Precondition: The parameter must be of type string to be relevant for the code.