[Cleanup] Remove unused usings
Escalated IDE0005 to warning for shipping code and executed FixAll on all unused usings in the solution. I retained the diagnostic as hidden diagnostic for test code as we have lot of autogenerated test files with helpful usings to help someone get started whenever they start authoring tests.

The diagnostic is still not enforced on CI as the analyzer is not yet part of IDE code style analyzer NuGet package.