Fix DFA assert for ??= operator
The CFG for ??= operator contains an LValue FlowCapture with a flow capture reference as both left of an assignment and value of another flow capture. The latter needs to be treated specially as a dereference to the flow capture.

Unit tests for this operator needs C# 8 and Roslyn 3.x, so I will add them in the 3.x branch.

Fixes #2782