Improve DFA performance
Add PointsToAnalysisKind enum to allow partial/complete points to analysis. Complete analysis mode is same as implemented today. Partial analysis mode skips tracking points to values for fields and properties, which is one of the primary cause of large points to data maps and slow DFA execution, and would be used by majority of non-security rules that do not need tracking fields/properties. Security rules need complete analysis and should be unaffected. We would still need to investigate in performance improvements for complete analysis mode, but the hope is this change should still improve the overall performance for the entire analyzers package and address all the performance reports for the same.