Add an analyzer for PublishSingleFile (#3921)
When publishing single-file applications, certain APIs are very likely
to return incorrect results. Most notably, asking for the Location of
any Assembly loaded from the single-file will return an empty string,
instead of a file path. To minimize surprise, this analyzer flags all
usages of the APIs most likely to break when used in single-file when
the PublishSingleFile property is set to true, and the
IncludeAllContentForSelfExtractProperty is not true (which provides the
legacy API behavior).

Co-authored-by: Amaury Levé <evangelink@gmail.com>
Co-authored-by: Manish Vasani <mavasani@microsoft.com>