Reduce rebuilding GenerateAnalyzerNuspec.csproj
PR #4442 has reports where the build fails because the project is being
rebuilt multiple time with highly parallel builds and that causes file
conflicts.

I am not sure why that's happening, to be honest. Or at least why it's
only happening for GenerateAnalyzerNuspec. That project uses the
`MSBuild` + `Exec` pattern:

https://github.com/dotnet/roslyn-analyzers/blob/37eba3a5aa2d706745f37692ad7ba7cb0a66a115/eng/GenerateAnalyzerNuspec.targets#L140-L144

That's the same pattern used elsewhere in the same file:

https://github.com/dotnet/roslyn-analyzers/blob/37eba3a5aa2d706745f37692ad7ba7cb0a66a115/eng/GenerateAnalyzerNuspec.targets#L105-L109

Still, this PR tries to reduce how many times that project is rebuilt,
which should (hopefully) reduce the frequency of the error, if not
eliminate it.

Any suggestions on what could be causing this specifically in
GenerateAnalyzerNuspec?