Remove AppHosts from tools
The tool projects are building the apphost, but we always invoke the
tool through `dotnet tool.dll` instead of the apphost (`tool`). Removing
the apphost should have no effts.

Keeping the AppHosts introduces additional prebuilts for source build
and it would be nice to not have to maintain this patch in source-build.
See https://github.com/dotnet/source-build/issues/1905 for more details.