Fix detection of duplicate symbols with oblivious markers
Previously, api file validation would only check for exact matches of symbols.
The addition of the oblivious marker for nullable and non-nullable annotations broke the duplication validation.
The change is to ignore the oblivious marker on duplicate comparison to prevent duplicates going unnoticed when adding the nullable reference type support.