Changes to support 'PVP flow' in source-build (#6454)
PVP flow is a method of flowing dependencies in source build that only overrides versions of packages that the repo has a declared dependency on in Version.Details.xml. This means that the package flow is closer to the Microsoft's way of building .NET.

To enable this for roslyn-analyzers, we just need to add dependencis on the packages that should be overridden in source build. In this case, it's the compiler toolset package, and the Microsoft.CodeAnalysis package. I also cleaned up the CodeAnalysisVersion property to match the package name, however, I generally avoided adding dependencies that do not participate in source build. For example, System.CommandLine is used only in tooling that is not included in SB. It may be worth adding these dependencies to make it easier to track and update dependencies.

To fully enable PVP flow, the repo project in the VMR will needto set `<PackageVersionPropsFlowType>DependenciesOnly</PackageVersionPropsFlowType>` once these changes flow in.