CA2263: Prefer generic overload when type is known
This analyzer detects when a `System.Type` overload is called when a
suitable generic overload is available.

To validate if a generic overload is applicable, the arity, parameter
count, containing symbol of the invocation (to avoid endless loops),
return type, argument types and type constraints(using speculative
binding) are checked.

The fixer removes unnecessary casts and parentheses.