CA1853: Address issues found in PR review #6767 (#6791)
* Rename do not guard set analyzer and fixer

This is done in a separate commit, so that the actual changes can be
seen in the next commit. This is because git would not recognize the
rename if the changes were added as is.

Note that this commit by itself is broken.

* Combine analyzers for guarded calls

This combines the analyzers CA1853 and CA1868 to avoid code duplication,
make CA1853 support the same range of cases as CA1868 and fix #6781.