Add CA1514: Remove redundant length argument (#6814)
* Add CA1514: Remove redundant length argument

This analyzer detects uses of Substring/Slice(start, length) that slice
to the end of the buffer. In these cases, the length check is redundant,
can be omitted, and the call can be replaced with
Substring/Slice(start). The analyzer also detects if the evaluation of
the instance or arguments can lead to side effects and does not flag
these cases.

* Rename LengthCheck to LengthCalculation