Add CA1872: Prefer 'Convert.ToHexString' over 'BitConverter.ToString'
This analyzer detects the usage of the pattern
`BitConverter.ToString(bytes).Replace("-", "")` to convert an array of
bytes to an uppercase hex string (without hyphens in between) and
replaces it with a call to `Convert.ToHexString(bytes)`.
The analyzer will also try to preserve chaining `ToLower*` in between
for a lowercase hex string.