Execute NameAndParametersValid only when needed.
The NameAndParametersValid is executed for ALL methods in a class,
independenlty of the methods being static and returning a boolean.

If you look carefully at the code, we have a || after a && therefore
the NamedAndParametersValid is executed ALWAYS. This means that we are
performing the check for EVEVERY method, rather than on those methods
that are static and return a bool.