Rework how expensive argument evaluation is detected
Previously, some specific operations were treated as non-expensive and
the rest were. This is not really future-proof as the language evolves,
and is generally more prone to false positives.

So this has now been changed to the opposite: Treat certain operations
as expensive, and the rest are not expensive by default and do not
trigger a diagnostic.